Linting

+++ src/curl_multi_wrapper.h
@@ -3,16 +3,19 @@
 #ifndef SRC_CURL_MULTI_WRAPPER_H_
 #define SRC_CURL_MULTI_WRAPPER_H_
 
+#include <map>
+
 #include "curl/curl.h"
 #include "node.h"
+
 #include "helpers.h"
-#include <map>
 #include "curl_easy_wrapper.h"
 
 namespace nodecurl {
   class CurlMultiWrapper : public node::ObjectWrap {
     public:
       static void Initialize(v8::Handle<v8::Object> target);
+
     private:
       typedef std::map<curl_socket_t, ev_io> SockFDs;
 
@@ -24,7 +27,7 @@ namespace nodecurl {
       static v8::Handle<v8::Value> Execute(const v8::Arguments& args);
       static v8::Handle<v8::Value> Close(const v8::Arguments& args);
 
-      static int TimerFunction(CURLM* mh, long timeout, void* userp);
+      static int TimerFunction(CURLM* mh, int timeout, void* userp);
       static int SocketFunction(CURLM* mh, curl_socket_t sockfd,
           int events, void* userp, void* socketp);
       static void IOEventFunction(EV_P_ ev_io* watcher, int events);
